+++ crates/ruma-state-res/src/events/power_levels.rs
@@ -11,7 +11,7 @@ use ruma_common::{
     room_version_rules::AuthorizationRules,
     serde::{
         btreemap_deserialize_v1_powerlevel_values, deserialize_v1_powerlevel, from_raw_json_value,
-        DebugAsRefStr, DisplayAsRefStr, JsonObject, OrdAsRefStr, PartialEqAsRefStr,
+        DebugAsRefStr, DisplayAsRefStr, EqAsRefStr, JsonObject, OrdAsRefStr,
     },
     OwnedUserId, UserId,
 };
@@ -327,7 +327,7 @@ impl<E: Event> RoomPowerLevelsEventOptionExt for Option<RoomPowerLevelsEvent<E>>
 }
 
 /// Fields in the `content` of an `m.room.power_levels` event with an integer value.
-#[derive(DebugAsRefStr, Clone, Copy, DisplayAsRefStr, PartialEqAsRefStr, Eq, OrdAsRefStr)]
+#[derive(DebugAsRefStr, Clone, Copy, DisplayAsRefStr, EqAsRefStr, OrdAsRefStr)]
 #[non_exhaustive]
 pub enum RoomPowerLevelsIntField {
     /// `users_default`